refactor(web): consolidate agent code into src/lib/agents/
Merge scattered agent helpers, API functions, and types into a single
canonical module at src/lib/agents/{types,svc,utils}.ts, eliminating
duplicated implementations across lib/agents.ts, app/admin/agents/lib.ts,
refresh-pages/admin/AgentsPage/svc.ts, and related files.
