--- web/src/app/admin/agents/lib.ts
@@ -1,245 +1,0 @@
-import {
-  MinimalPersonaSnapshot,
-  Persona,
-  StarterMessage,
-} from "@/app/admin/agents/interfaces";
-
-interface PersonaUpsertRequest {
-  name: string;
-  description: string;
-  system_prompt: string;
-  task_prompt: string;
-  datetime_aware: boolean;
-  document_set_ids: number[];
-  is_public: boolean;
-  default_model_configuration_id: number | null;
-  starter_messages: StarterMessage[] | null;
-  users?: string[];
-  groups: number[];
-  tool_ids: number[];
-  remove_image?: boolean;
-  uploaded_image_id: string | null;
-  icon_name: string | null;
-  search_start_date: Date | null;
-  is_featured: boolean;
-  display_priority: number | null;
-  label_ids: number[] | null;
-  user_file_ids: string[] | null;
-  replace_base_system_prompt: boolean;
-  // Hierarchy nodes (folders, spaces, channels) for scoped search
-  hierarchy_node_ids: number[];
-  // Individual documents for scoped search
-  document_ids: string[];
-}
-
-export interface PersonaUpsertParameters {
-  name: string;
-  description: string;
-  system_prompt: string;
-  replace_base_system_prompt: boolean;
-  task_prompt: string;
-  datetime_aware: boolean;
-  document_set_ids: number[];
-  is_public: boolean;
-  default_model_configuration_id?: number | null;
-  starter_messages: StarterMessage[] | null;
-  users?: string[];
-  groups: number[];
-  tool_ids: number[];
-  remove_image?: boolean;
-  search_start_date: Date | null;
-  uploaded_image_id: string | null;
-  icon_name: string | null;
-  is_featured: boolean;
-  label_ids: number[] | null;
-  user_file_ids: string[];
-  // Hierarchy nodes (folders, spaces, channels) for scoped search
-  hierarchy_node_ids?: number[];
-  // Individual documents for scoped search
-  document_ids?: string[];
-}
-
-function buildPersonaUpsertRequest({
-  name,
-  description,
-  system_prompt,
-  task_prompt,
-  document_set_ids,
-  is_public,
-  groups,
-  datetime_aware,
-  users,
-  tool_ids,
-  remove_image,
-  search_start_date,
-  user_file_ids,
-  hierarchy_node_ids,
-  document_ids,
-  icon_name,
-  uploaded_image_id,
-  is_featured,
-  default_model_configuration_id,
-  starter_messages,
-  label_ids,
-  replace_base_system_prompt,
-}: PersonaUpsertParameters): PersonaUpsertRequest {
-  return {
-    name,
-    description,
-    system_prompt,
-    task_prompt,
-    document_set_ids,
-    is_public,
-    uploaded_image_id,
-    icon_name,
-    groups,
-    users,
-    tool_ids,
-    remove_image,
-    search_start_date,
-    datetime_aware,
-    is_featured: is_featured ?? false,
-    default_model_configuration_id: default_model_configuration_id ?? null,
-    starter_messages: starter_messages ?? null,
-    display_priority: null,
-    label_ids: label_ids ?? null,
-    user_file_ids: user_file_ids ?? null,
-    replace_base_system_prompt,
-    hierarchy_node_ids: hierarchy_node_ids ?? [],
-    document_ids: document_ids ?? [],
-  };
-}
-
-export async function uploadFile(file: File): Promise<string | null> {
-  const formData = new FormData();
-  formData.append("file", file);
-  const response = await fetch("/api/admin/persona/upload-image", {
-    method: "POST",
-    body: formData,
-    credentials: "include",
-  });
-
-  if (!response.ok) {
-    console.error("Failed to upload file");
-    return null;
-  }
-
-  const responseJson = await response.json();
-  return responseJson.file_id;
-}
-
-export async function createPersona(
-  personaUpsertParams: PersonaUpsertParameters
-): Promise<Response | null> {
-  const createPersonaResponse = await fetch("/api/persona", {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify(buildPersonaUpsertRequest(personaUpsertParams)),
-    credentials: "include",
-  });
-
-  return createPersonaResponse;
-}
-
-export async function updatePersona(
-  id: number,
-  personaUpsertParams: PersonaUpsertParameters
-): Promise<Response | null> {
-  const updatePersonaResponse = await fetch(`/api/persona/${id}`, {
-    method: "PATCH",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify(buildPersonaUpsertRequest(personaUpsertParams)),
-    credentials: "include",
-  });
-
-  return updatePersonaResponse;
-}
-
-export function deletePersona(personaId: number) {
-  return fetch(`/api/persona/${personaId}`, {
-    method: "DELETE",
-    credentials: "include",
-  });
-}
-
-function smallerNumberFirstComparator(a: number, b: number) {
-  return a > b ? 1 : -1;
-}
-
-function closerToZeroNegativesFirstComparator(a: number, b: number) {
-  if (a < 0 && b > 0) {
-    return -1;
-  }
-  if (a > 0 && b < 0) {
-    return 1;
-  }
-
-  const absA = Math.abs(a);
-  const absB = Math.abs(b);
-
-  if (absA === absB) {
-    return a > b ? 1 : -1;
-  }
-
-  return absA > absB ? 1 : -1;
-}
-
-export function personaComparator(
-  a: MinimalPersonaSnapshot | Persona,
-  b: MinimalPersonaSnapshot | Persona
-) {
-  if (a.display_priority === null && b.display_priority === null) {
-    return closerToZeroNegativesFirstComparator(a.id, b.id);
-  }
-
-  if (a.display_priority !== b.display_priority) {
-    if (a.display_priority === null) {
-      return 1;
-    }
-    if (b.display_priority === null) {
-      return -1;
-    }
-
-    return smallerNumberFirstComparator(a.display_priority, b.display_priority);
-  }
-
-  return closerToZeroNegativesFirstComparator(a.id, b.id);
-}
-
-export async function togglePersonaFeatured(
-  personaId: number,
-  featured: boolean
-) {
-  const response = await fetch(`/api/admin/persona/${personaId}/featured`, {
-    method: "PATCH",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify({
-      is_featured: !featured,
-    }),
-    credentials: "include",
-  });
-  return response;
-}
-
-export async function togglePersonaVisibility(
-  personaId: number,
-  isVisible: boolean
-) {
-  const response = await fetch(`/api/admin/persona/${personaId}/listed`, {
-    method: "PATCH",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify({
-      is_listed: !isVisible,
-    }),
-    credentials: "include",
-  });
-  return response;
-}
