+++ web/src/app/admin/bots/[bot-id]/channels/SlackChannelConfigFormFields.tsx
@@ -12,7 +12,7 @@ import {
   TextFormField,
 } from "@/components/Field";
 import { Button, Divider } from "@opal/components";
-import { MinimalPersonaSnapshot } from "@/app/admin/agents/interfaces";
+import { MinimalPersonaSnapshot } from "@/lib/agents/types";
 import DocumentSetCard from "@/sections/cards/DocumentSetCard";
 import CollapsibleSection from "@/app/admin/agents/CollapsibleSection";
 import { StandardAnswerCategoryResponse } from "@/components/standardAnswers/getStandardAnswerCategoriesIfEE";
