+++ web/src/app/admin/discord-bot/[guild-id]/DiscordChannelsTable.tsx
@@ -19,7 +19,7 @@ import {
 } from "@/app/admin/discord-bot/types";
 import { SvgHash, SvgBubbleText, SvgLock } from "@opal/icons";
 import { IconProps } from "@opal/types";
-import { Persona } from "@/app/admin/agents/interfaces";
+import { Persona } from "@/lib/agents/types";
 
 function getChannelIcon(
   channelType: DiscordChannelType,
