+++ web/src/app/app/components/files/images/utils.ts
@@ -5,10 +5,6 @@ export function buildImgUrl(fileId: string) {
   return `/api/chat/file/${fileId}`;
 }
 
-export function buildAgentAvatarUrl(agentId: number) {
-  return `/api/persona/${agentId}/avatar`;
-}
-
 /**
  * If `href` points to a chat file and `linkText` ends with an image extension,
  * returns the file ID. Otherwise returns null.
