+++ web/src/app/ee/admin/performance/usage/PersonaMessagesChart.tsx
@@ -18,7 +18,7 @@ import {
   SelectValue,
 } from "@/components/ui/select";
 import { useState, useMemo, useEffect } from "react";
-import { Persona } from "@/app/admin/agents/interfaces";
+import { Persona } from "@/lib/agents/types";
 
 export function PersonaMessagesChart({
   availablePersonas,
