+++ web/src/lib/agents/utils.ts
@@ -1,0 +1,75 @@
+import { User } from "@/lib/types";
+import { checkUserIsNoAuthUser } from "@/lib/user";
+import { MinimalPersonaSnapshot, Persona } from "@/lib/agents/types";
+
+export function checkUserOwnsAgent(
+  user: User | null,
+  agent: MinimalPersonaSnapshot | Persona
+) {
+  return checkUserIdOwnsAgent(user?.id, agent);
+}
+
+export function checkUserIdOwnsAgent(
+  userId: string | undefined,
+  agent: MinimalPersonaSnapshot | Persona
+) {
+  return (
+    !!userId &&
+    (checkUserIsNoAuthUser(userId) || agent.owner?.id === userId) &&
+    !agent.builtin_persona
+  );
+}
+
+function smallerNumberFirstComparator(a: number, b: number) {
+  return a > b ? 1 : -1;
+}
+
+function closerToZeroNegativesFirstComparator(a: number, b: number) {
+  if (a < 0 && b > 0) {
+    return -1;
+  }
+  if (a > 0 && b < 0) {
+    return 1;
+  }
+
+  const absA = Math.abs(a);
+  const absB = Math.abs(b);
+
+  if (absA === absB) {
+    return a > b ? 1 : -1;
+  }
+
+  return absA > absB ? 1 : -1;
+}
+
+export function personaComparator(
+  a: MinimalPersonaSnapshot | Persona,
+  b: MinimalPersonaSnapshot | Persona
+) {
+  if (a.display_priority === null && b.display_priority === null) {
+    return closerToZeroNegativesFirstComparator(a.id, b.id);
+  }
+
+  if (a.display_priority !== b.display_priority) {
+    if (a.display_priority === null) {
+      return 1;
+    }
+    if (b.display_priority === null) {
+      return -1;
+    }
+
+    return smallerNumberFirstComparator(a.display_priority, b.display_priority);
+  }
+
+  return closerToZeroNegativesFirstComparator(a.id, b.id);
+}
+
+export function filterAgents(
+  assistants: MinimalPersonaSnapshot[]
+): MinimalPersonaSnapshot[] {
+  return assistants.filter((a) => a.is_listed).sort(personaComparator);
+}
+
+export function buildAgentAvatarUrl(agentId: number) {
+  return `/api/persona/${agentId}/avatar`;
+}
