+++ web/src/refresh-pages/admin/AgentsPage/AgentsTable.tsx
@@ -9,13 +9,13 @@ import Text from "@/refresh-components/texts/Text";
 import InputTypeIn from "@/refresh-components/inputs/InputTypeIn";
 import type { MinimalUserSnapshot } from "@/lib/types";
 import AgentAvatar from "@/refresh-components/avatars/AgentAvatar";
-import type { MinimalPersonaSnapshot } from "@/app/admin/agents/interfaces";
+import type { MinimalPersonaSnapshot } from "@/lib/agents/types";
 import { useAdminAgents } from "@/hooks/useAgents";
 import { toast } from "@/hooks/useToast";
 import AgentRowActions from "@/refresh-pages/admin/AgentsPage/AgentRowActions";
-import { updateAgentDisplayPriorities } from "@/refresh-pages/admin/AgentsPage/svc";
-import type { AgentRow } from "@/refresh-pages/admin/AgentsPage/interfaces";
-import type { Persona } from "@/app/admin/agents/interfaces";
+import { updateAgentDisplayPriorities } from "@/lib/agents/svc";
+import type { AgentRow } from "@/lib/agents/types";
+import type { Persona } from "@/lib/agents/types";
 import { SvgUser } from "@opal/icons";
 import { DEFAULT_PAGE_SIZE } from "@/lib/constants";
 import { Section } from "@/layouts/general-layouts";
